Added Output Mailer for mail notifications
